Fix blobutils_get_int32
git-svn-id: https://svn.mcs.anl.gov/repos/exm/sfw/turbine/trunk@9501 dc4e9af1-7f46-4ead-bba6-71afc04862de

+++ code/src/tcl/blob/blob.c
@@ -230,10 +230,10 @@ blobutils_get_int(int* pointer, int index)
   return pointer[index];
 }
 
-int32_t
+int
 blobutils_get_int32(int32_t* pointer, int index)
 {
-  return pointer[index];
+  return (int) pointer[index];
 }
 
 /**
